packages/forge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "animus-forge"
version = "1.2.0"
description = "Multi-agent orchestration framework for production AI workflows"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "AreteDriver", email = "aretedriver@gmail.com"}
]
keywords = ["multi-agent", "ai", "orchestration", "workflow"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=2.15.0",
    "anthropic>=0.79.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "streamlit>=1.53.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.3.0",
    "google-api-python-client>=2.150.0",
    "notion-client>=3.0.0",
    "PyGithub>=2.4.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=25.1.0",
    "apscheduler>=3.11.0",
    "urllib3>=2.6.0",
    "pynacl>=1.6.0",
    "pyasn1>=0.6.1",
    "typer>=0.23.0",
    "rich>=14.0.0",
    "pyyaml>=6.0.0",
    "jsonschema>=4.23.0",
    "slowapi>=0.1.9",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "textual>=8.0.0",
    "convergentai>=1.1.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
mcp = ["mcp>=1.0.0"]
messaging = ["python-telegram-bot>=22.0", "discord-py>=2.3.0"]
browser = ["playwright>=1.40.0"]
local = ["httpx>=0.28.0", "psutil>=7.0.0"]
all = [
    "animus-forge[postgres,mcp,messaging,browser,local]",
]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.15.1",
]

[project.scripts]
animus-forge = "animus_forge.cli:app"

[project.urls]
Homepage = "https://github.com/AreteDriver/animus"
Repository = "https://github.com/AreteDriver/animus"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/animus_forge"]
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__",
]
